Added stage to generate multiple compressed textures
