+++ lib/Pipeline.js
@@ -11,8 +11,9 @@ var bakeAmbientOcclusion = require('./bakeAmbientOcclusion');
 var combineMeshes = require('./combineMeshes');
 var combineNodes = require('./combineNodes');
 var compressIntegerAccessors = require('./compressIntegerAccessors');
-var compressTextures = require('./compressTextures');
 var compressTextureCoordinates = require('./compressTextureCoordinates');
+var compressTextures = require('./compressTextures');
+var compressTexturesMultipleFormats = require('./compressTexturesMultipleFormats');
 var combinePrimitives = require('./combinePrimitives');
 var convertDagToTree = require('./convertDagToTree');
 var encodeImages = require('./encodeImages');
@@ -34,6 +35,7 @@ var writeSource = require('./writeSource');
 
 var defaultValue = Cesium.defaultValue;
 var defined = Cesium.defined;
+var isArray = Cesium.isArray;
 
 module.exports = Pipeline;
 
@@ -80,7 +82,7 @@ Pipeline.processJSON = function(gltf, options) {
  * @param {Object} [options.compressTextureCoordinates=false] Flag to run compressTextureCoordinates stage.
  * @param {Object} [options.kmcOptions=undefined] Options to pass to the generateModelMaterialsCommon stage, if undefined, stage is not run.
  * @param {Object} [options.quantize] Flag to run quantizeAttributes stage.
- * @param {Object} [options.textureCompressionOptions=undefined] Options to pass to the compressTextures stage, if undefined, stage is not run.
+ * @param {Object|Object[]} [options.textureCompressionOptions=undefined] Options to pass to the compressTextures stage. If an array of options is given, the textures will be compressed in multiple formats. If undefined, stage is not run.
  * @returns {Promise} A promise that resolves to the processed glTF asset.
  */
 Pipeline.processJSONWithExtras  = function(gltfWithExtras, options) {
@@ -116,7 +118,7 @@ Pipeline.processJSONWithExtras  = function(gltfWithExtras, options) {
     }
     
     var aoOptions = options.aoOptions;
-    if (defined(aoOptions) && aoOptions.enable) {
+    if (defined(aoOptions)) {
         bakeAmbientOcclusion(gltfWithExtras, aoOptions);
     }
 
@@ -146,15 +148,19 @@ Pipeline.processJSONWithExtras  = function(gltfWithExtras, options) {
                 quantizeAttributes(gltfWithExtras, quantizedOptions);
             }
             var textureCompressionOptions = options.textureCompressionOptions;
-            if (defined(textureCompressionOptions) && textureCompressionOptions.enable) {
-                return compressTextures(gltfWithExtras, textureCompressionOptions);
+            if (defined(textureCompressionOptions)) {
+                if (isArray(textureCompressionOptions)) {
+                    return compressTexturesMultipleFormats(gltfWithExtras, textureCompressionOptions);
+                } else {
+                    return compressTextures(gltfWithExtras, textureCompressionOptions);
+                }
             } else {
                 return encodeImages(gltfWithExtras);
             }
         })
         .then(function() {
             var kmcOptions = options.kmcOptions;
-            if (defined(kmcOptions) && kmcOptions.enable) {
+            if (defined(kmcOptions)) {
                 generateModelMaterialsCommon(gltfWithExtras, options.kmcOptions);
             }
             return gltfWithExtras;
