+++ lib/parseArguments.js
@@ -1,10 +1,12 @@
 'use strict';
 var Cesium = require('cesium');
+var clone = require('clone');
 var path = require('path');
 var yargs = require('yargs');
 
 var defaultValue = Cesium.defaultValue;
 var defined = Cesium.defined;
+var isArray = Cesium.isArray;
 
 module.exports = parseArguments;
 
@@ -144,7 +146,7 @@ function parseArguments(args) {
             },
             'texcomp.format': {
                 choices: ['pvrtc1', 'pvrtc2', 'etc1', 'etc2', 'astc', 'dxt1', 'dxt3', 'dxt5', 'crunch-dxt1', 'crunch-dxt5'],
-                describe: 'The compressed texture format.',
+                describe: 'The compressed texture format. Multiple formats may be supplied by repeating this flag.',
                 group: 'Options: Texture Compression',
                 type: 'string'
             },
@@ -177,18 +179,35 @@ function parseArguments(args) {
         }).parse(args);
 
     // If any raw ao, kmc, or texcomp parameters were specified, they are enabled
+    var i;
+    var aoOptions;
+    var kmcOptions;
+    var textureCompressionOptions;
     var nargs = process.argv.length;
-    for (var i = 0; i < nargs; i++) {
+    for (i = 0; i < nargs; i++) {
         var arg = process.argv[i];
         if (arg.indexOf('ao') >= 0) {
-            argv.ao.enable = true;
+            aoOptions = argv.ao;
         }
         if (arg.indexOf('kmc') >= 0) {
-            argv.kmc.enable = true;
+            kmcOptions = argv.kmc;
         }
         if (arg.indexOf('texcomp') >= 0) {
-            argv.texcomp.enable = true;
+            textureCompressionOptions = argv.texcomp;
         }
+    }
+
+    // If multiple compressed texture formats are specified send in an array of options
+    if (defined(textureCompressionOptions) && isArray(textureCompressionOptions.format)) {
+        var formats = textureCompressionOptions.format;
+        var length = formats.length;
+        var optionsArray = new Array(length);
+        for (i = 0; i < length; ++i) {
+            var options = clone(textureCompressionOptions);
+            options.format = formats[i];
+            optionsArray[i] = options;
+        }
+        textureCompressionOptions = optionsArray;
     }
 
     var gltfPath = defaultValue(argv.i, argv._[0]);
@@ -214,7 +233,7 @@ function parseArguments(args) {
     }
 
     return {
-        aoOptions: argv.ao,
+        aoOptions: aoOptions,
         binary: argv.b,
         compressTextureCoordinates: argv.c,
         embed: !argv.s,
@@ -223,12 +242,12 @@ function parseArguments(args) {
         faceNormals: argv.f,
         tangentsBitangents : argv.tangentsBitangents,
         inputPath: gltfPath,
-        kmcOptions: argv.kmc,
+        kmcOptions: kmcOptions,
         optimizeForCesium: argv.cesium,
         outputPath: outputPath,
         preserve: argv.p,
         quantize: argv.q,
         removeNormals: argv.r,
-        textureCompressionOptions: argv.texcomp
+        textureCompressionOptions: textureCompressionOptions
     };
 }
